+++ beam-core/Database/Beam/Query/CustomSQL.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE FunctionalDependencies #-}
 {-# LANGUAGE UndecidableInstances #-}
-{-# LANGUAGE CPP #-}
 
 -- | Allows the creation of custom SQL expressions from arbitrary string-like values.
 --