+++ beam-migrate/Database/Beam/Migrate/SQL/SQL92.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE CPP #-}
 
 -- | Finally-tagless encoding of SQL92 DDL commands.
 --