+++ beam-migrate/Database/Beam/Migrate/Serialization.hs
@@ -1,6 +1,5 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE TupleSections #-}
-{-# LANGUAGE CPP #-}
 
 -- | Serialization and deserialization helpers for beam data types.
 --
@@ -40,9 +39,7 @@ import           Control.Applicative
 import           Control.Monad
 
 import           Data.Aeson
-#if MIN_VERSION_aeson(2,0,0)
 import qualified Data.Aeson.Key as DAK
-#endif
 import           Data.Aeson.Types (Parser)
 import qualified Data.Dependent.Map as D
 import qualified Data.GADT.Compare as D
@@ -331,13 +328,8 @@ sql92Deserializers = mconcat
                    , beamDeserializer deserializeSql92ReferentialAction
                    , beamDeserializer deserializeSql92Attributes ]
   where
-#if MIN_VERSION_aeson(2,0,0)
-    makeKey = DAK.fromText
-#else
-    makeKey = id
-#endif
     parseSub nm o key parse =
-      withObject (unpack (nm <> "." <> key)) parse =<< o .: makeKey key
+      withObject (unpack (nm <> "." <> key)) parse =<< o .: DAK.fromText key
 
     deserializeSql92DataType :: BeamDeserializers be' -> Value
                              -> Parser (BeamSqlBackendDataTypeSyntax be)