+++ beam-sqlite/Database/Beam/Sqlite/Syntax.hs
@@ -5,7 +5,6 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE CPP #-}
 
 -- | SQLite implementations of the Beam SQL syntax classes
 --